Trójki i czwórki są sprawdzone. Trzeba sprawdzić pary.

+++ warsztat1/zadania_pwr2-master/interfaces/sum_item/sum_item.h
@@ -4,12 +4,12 @@
 #define SUM_ITEM_H
 
 struct SumItem {
-    int sum;
-    int firstElementIndex;
-    int secondElementIndex;
+    long long int sum;
+    size_t firstElementIndex;
+    size_t secondElementIndex;
 };
 
-struct SumItem* createSumItem(int sum, int firstElementIndex, int secondElementIndex);
+struct SumItem* createSumItem(long long int sum, size_t firstElementIndex, size_t secondElementIndex);
 bool checkIfSumItemsHaveCommonElements(struct SumItem* firstItem, struct SumItem* secondItem);
 
 #endif
